Add files via upload
I added Monster Name, lvl, and health bar. 

next thing to add is  probably the animation per attack... but thats going to be a lot of work. i'm putting this on my website so i'm probably only going to focus on pikachu's move sets for now. ;) 

I also added a secret easter egg encounter ;D. Obviously not done but its just for fun. for now. :D